% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpgi_annot.R
\name{cpgi_annot}
\alias{cpgi_annot}
\title{CPGI Status annotation}
\usage{
cpgi_annot(probe_ids, method = c("chisq", "fisher"), arch = "hm450",
  genome = "hg19", soi_id = NULL)
}
\arguments{
\item{probe_ids}{A character vector containing the Illumina probe identifiers
to be tested.}

\item{method}{A character scalar indicating the testing method.}

\item{arch}{A character scalar indicating the array architecture.}

\item{genome}{A character scalar indicating the reference genome.}

\item{soi_id}{A character scalar describing the subset of interest being
tested. A value of NULL does not even add a column with the identifier.}
}
\value{
A data.frame containing the test results.
}
\description{
CPGI Status annotation
}
