% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cpgi_hm450_hg19}
\alias{cpgi_hm450_hg19}
\title{CPGI Status for Infinium450k in hg19.}
\format{A data frame with 591542 rows and 2 variables:
\describe{
  \item{Probe_ID}{Illumina probe identifier}
  \item{CPGI_Status}{CPGI Status}
}}
\usage{
cpgi_hm450_hg19
}
\description{
A dataset containing the CPGI Status for every probe in the Illumina
Infinium 450k. Probes can belong to multiple categories.
}
\keyword{datasets}
