% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cpgi.R
\name{run_cpgi}
\alias{run_cpgi}
\title{Run CPGI Annotation interactive tool.}
\usage{
run_cpgi()
}
\value{
This function normally does not return unless interrupted.
}
\description{
Run an interactive Shiny tool providing a user interface to the cpgi_annot()
function. Thus, a user without experience may annotate a given set of probe
identifiers without effort.
}
